pyproject.toml
[tool.poetry]
name = "claude-clinical-bridge"
version = "0.1.0"
description = "AI-powered healthcare clinical documentation and coding assistant using Claude Agent SDK and MCP"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
# MCP and Claude Agent SDK
mcp = ">=1.2.0"
anthropic = "^0.40.0"

# FHIR Integration
fhirclient = "^4.2.1"
fhir-resources = "^7.1.0"

# Vector Database and Embeddings
qdrant-client = "^1.11.3"
sentence-transformers = "^3.3.1"

# Configuration and Logging
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
structlog = "^24.4.0"
python-dotenv = "^1.0.1"

# API Framework (for future TypeScript integration)
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}

# Security and Encryption
cryptography = "^44.0.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}

# Async and HTTP
httpx = "^0.28.1"
aiofiles = "^24.1.0"

# Data Processing
pandas = "^2.2.3"
numpy = "^2.2.1"

# Retry Logic
tenacity = "^9.0.0"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"

# Linting and Type Checking
ruff = "^0.8.4"
mypy = "^1.14.0"
types-aiofiles = "^24.1.0"

# Security Scanning
bandit = {extras = ["toml"], version = "^1.8.0"}
safety = "^3.2.11"

# Development Tools
ipython = "^8.31.0"
ipdb = "^0.13.13"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "fhirclient.*",
    "sentence_transformers.*",
    "qdrant_client.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "venv"]
skips = ["B101", "B601"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
